+++ test/responses/occupancySensor.ts
@@ -1,0 +1,63 @@
+import { OccupancySensor } from '../../src/types/device/OccupancySensor'
+
+export const occupancySensor: OccupancySensor = {
+  id: '00000000-0000-0000-0000-000000000000',
+  relationId: '00000000-0000-0000-0000-000000000000',
+  type: 'sensor',
+  deviceType: 'occupancySensor',
+  createdAt: '2000-01-01T00:00:00.000Z',
+  isReachable: true,
+  lastSeen: '2000-01-01T00:00:00.000Z',
+  customIcon: 'products_matter_wrlss_mtn_sensor',
+  attributes: {
+    identifyPeriod: 0,
+    identifyStarted: '2000-01-01T00:00:00.000Z',
+    isDetected: false,
+    customName: 'MYGGSPRAY wrlss mtn sensor',
+    firmwareVersion: '1.0.7',
+    hardwareVersion: 'P2.0',
+    manufacturer: 'IKEA of Sweden',
+    model: 'MYGGSPRAY wrlss mtn sensor',
+    productCode: 'E2494',
+    serialNumber: 'UNKNOWN_SERIAL_NUMBER',
+    discriminator: -1,
+    permittingJoin: false,
+    qrCode: '',
+    setupCode: '',
+    windowOpen: false,
+    otaPolicy: 'autoUpdate',
+    otaProgress: 0,
+    otaScheduleEnd: '00:00',
+    otaScheduleStart: '00:00',
+    otaState: 'readyToCheck',
+    otaStatus: 'upToDate',
+    batteryPercentage: 86,
+    sensorConfig: {
+      scheduleOn: true,
+      onDuration: 0,
+      schedule: {
+        onCondition: {
+          time: 'sunset',
+          offset: 45,
+        },
+        offCondition: {
+          time: 'sunrise',
+          offset: 45,
+        },
+      },
+    },
+  },
+  capabilities: {
+    canSend: [],
+    canReceive: ['customName', 'windowOpen'],
+  },
+  room: {
+    id: '00000000-0000-0000-0000-000000000000',
+    name: 'Room',
+    color: 'ikea_green_no_65',
+    icon: 'rooms_arm_chair',
+  },
+  deviceSet: [],
+  remoteLinks: [],
+  isHidden: false,
+}