+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/lpgera/dirigera/compare/v1.6.0...HEAD)
 
 - Add support for the BILRESA dual button remote, including blind and shortcut control modes.
+- Add MYGGSPRAY occupancy sensor and light sensor support.
 
 ## [1.6.0](https://github.com/lpgera/dirigera/compare/v1.5.0...v1.6.0) - 2025-12-23